+++ internal/service/loadbalancer/planmodifier.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 )
@@ -51,40 +52,118 @@ func validateNetworks(networkModels []loadbalancerNetworkModel) diag.Diagnostics
 	return diags
 }
 
-const networkRequiresReplaceIfDescription = "clearing the network field is only allowed when equivalent `networks` block is defined"
+func networkReplacedWithNetworks(ctx context.Context, state, plan loadBalancerModel) (bool, diag.Diagnostics) {
+	var diags diag.Diagnostics
 
-func networkRequiresReplaceIfFunc(ctx context.Context, req planmodifier.StringRequest, resp *stringplanmodifier.RequiresReplaceIfFuncResponse) {
-	prev := req.StateValue.ValueString()
-	next := req.PlanValue.ValueString()
-
-	if prev == "" || next != "" {
-		resp.RequiresReplace = true
-		return
+	// Ensure that `networks` value was null
+	if !state.Networks.IsNull() {
+		return false, diags
 	}
 
-	var plan loadBalancerModel
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+	// Ensure that `network` value is cleared
+	prevNetwork := state.Network.ValueString()
+	nextNetwork := plan.Network.ValueString()
 
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	if plan.Networks.IsNull() {
-		resp.RequiresReplace = true
-		return
+	if prevNetwork == "" || nextNetwork != "" {
+		return false, diags
 	}
 
 	var networkModels []loadbalancerNetworkModel
-	resp.Diagnostics.Append(plan.Networks.ElementsAs(ctx, &networkModels, false)...)
+	diags.Append(plan.Networks.ElementsAs(ctx, &networkModels, false)...)
+	if diags.HasError() {
+		return false, diags
+	}
+
+	// Equivalent `networks` block contains two entries. First for public network and second for private network where UUID should match network value.
+	if len(networkModels) != 2 {
+		return false, diags
+	}
+
+	if networkModels[0].Type.ValueString() != string(upcloud.LoadBalancerNetworkTypePublic) {
+		return false, diags
+	}
+
+	if networkModels[1].Type.ValueString() != string(upcloud.LoadBalancerNetworkTypePrivate) {
+		return false, diags
+	}
+
+	if networkModels[1].Network.ValueString() != prevNetwork {
+		return false, diags
+	}
+
+	return true, diags
+}
+
+const networkRequiresReplaceIfDescription = "clearing the `network` field is only allowed when equivalent `networks` block is defined"
+
+func networkRequiresReplaceIfFunc(ctx context.Context, req planmodifier.StringRequest, resp *stringplanmodifier.RequiresReplaceIfFuncResponse) {
+	var plan, state loadBalancerModel
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
 
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	// Equivalent `networks` block contains two entries. First for public network and second for private network where UUID should match network value.
-	if len(networkModels) != 2 || networkModels[1].ID.ValueString() != prev {
-		resp.RequiresReplace = true
+	replaced, diags := networkReplacedWithNetworks(ctx, state, plan)
+	resp.RequiresReplace = !replaced
+	resp.Diagnostics.Append(diags...)
+}
+
+const networksRequiresReplaceIfDescription = "modifying `networks` field is only allowed when it replaces equivalent `network` value"
+
+func networksRequiresReplaceIfFunc(ctx context.Context, req planmodifier.ListRequest, resp *listplanmodifier.RequiresReplaceIfFuncResponse) {
+	var plan, state loadBalancerModel
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+
+	if resp.Diagnostics.HasError() {
+		return
 	}
 
-	resp.RequiresReplace = false
+	replaced, diags := networkReplacedWithNetworks(ctx, state, plan)
+	resp.Diagnostics.Append(diags...)
+
+	if replaced {
+		resp.RequiresReplace = false
+		return
+	}
+
+	stateLen := len(state.Networks.Elements())
+	planLen := len(plan.Networks.Elements())
+
+	if stateLen == planLen {
+		// No change in length, nested fields have their own plan modifiers to require replace if needed.
+		resp.RequiresReplace = false
+		return
+	}
+
+	resp.RequiresReplace = true
+}
+
+func networksNestedValueRequiresReplaceIf() planmodifier.String {
+	return stringplanmodifier.RequiresReplaceIf(
+		func(ctx context.Context, req planmodifier.StringRequest, resp *stringplanmodifier.RequiresReplaceIfFuncResponse) {
+			var plan, state loadBalancerModel
+			resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+			resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+
+			if resp.Diagnostics.HasError() {
+				return
+			}
+
+			replaced, diags := networkReplacedWithNetworks(ctx, state, plan)
+			resp.Diagnostics.Append(diags...)
+
+			// Do not require replace when value changes from null to non-null value when `network` is replaced with `networks`
+			if replaced && req.StateValue.ValueString() == "" && req.PlanValue.ValueString() != "" {
+				resp.RequiresReplace = false
+				return
+			} else {
+				resp.RequiresReplace = true
+			}
+		},
+		networksRequiresReplaceIfDescription,
+		networksRequiresReplaceIfDescription,
+	)
 }