+++ upcloud/resource_upcloud_loadbalancer_test.go
@@ -345,19 +345,46 @@ func TestAccUpcloudLoadBalancer_minimal(t *testing.T) {
 }
 
 func TestAccUpcloudLoadBalancer_network(t *testing.T) {
-	testData := utils.ReadTestDataFile(t, "testdata/upcloud_loadbalancer/loadbalancer_network.tf")
+	testDataS1 := utils.ReadTestDataFile(t, "testdata/upcloud_loadbalancer/loadbalancer_network_s1.tf")
+	testDataS2 := utils.ReadTestDataFile(t, "testdata/upcloud_loadbalancer/loadbalancer_network_s2.tf")
+	testDataS3 := utils.ReadTestDataFile(t, "testdata/upcloud_loadbalancer/loadbalancer_network_s3.tf")
 
-	name := "upcloud_loadbalancer.this"
+	migrateName := "upcloud_loadbalancer.migrate_then_rename"
+	renameName := "upcloud_loadbalancer.migrate_and_rename"
+
+	var migrateUUID, renameUUID string
 
 	resource.ParallelTest(t, resource.TestCase{
 		PreCheck:                 func() { TestAccPreCheck(t) },
 		ProtoV6ProviderFactories: TestAccProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: testData,
+				Config: testDataS1,
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(name, "operational_state", "running"),
-					resource.TestCheckResourceAttr(name, "networks.#", "0"),
+					CheckStringDoesNotChange(migrateName, "id", &migrateUUID),
+					resource.TestCheckResourceAttr(migrateName, "operational_state", "running"),
+					resource.TestCheckResourceAttr(migrateName, "networks.#", "0"),
+					CheckStringDoesNotChange(renameName, "id", &renameUUID),
+					resource.TestCheckResourceAttr(renameName, "operational_state", "running"),
+					resource.TestCheckResourceAttr(renameName, "networks.#", "0"),
+				),
+			},
+			{
+				Config: testDataS2,
+				Check: resource.ComposeTestCheckFunc(
+					CheckStringDoesNotChange(migrateName, "id", &migrateUUID),
+					resource.TestCheckResourceAttr(migrateName, "operational_state", "running"),
+					resource.TestCheckResourceAttr(migrateName, "networks.#", "2"),
+					CheckStringDoesNotChange(renameName, "id", &renameUUID),
+					resource.TestCheckResourceAttr(renameName, "operational_state", "running"),
+					resource.TestCheckResourceAttr(renameName, "networks.#", "2"),
+				),
+			},
+			{
+				Config: testDataS3,
+				Check: resource.ComposeTestCheckFunc(
+					CheckStringDoesNotChange(migrateName, "id", &migrateUUID),
+					CheckStringDoesNotChange(renameName, "id", &renameUUID),
 				),
 			},
 		},