+++ upcloud/testdata/upcloud_loadbalancer/loadbalancer_network_s1.tf
@@ -1,6 +1,6 @@
 variable "basename" {
   type    = string
-  default = "tf-acc-test-lb-network-"
+  default = "tf-acc-test-lb-network-migrate-"
 }
 
 variable "zone" {
@@ -14,13 +14,21 @@ resource "upcloud_network" "this" {
 
   ip_network {
     address = "10.0.8.0/24"
-    dhcp = true
-    family = "IPv4"
+    dhcp    = true
+    family  = "IPv4"
   }
 }
 
-resource "upcloud_loadbalancer" "this" {
+resource "upcloud_loadbalancer" "migrate_then_rename" {
   name = "${var.basename}lb"
+  plan = "development"
+  zone = var.zone
+
+  network = upcloud_network.this.id
+}
+
+resource "upcloud_loadbalancer" "migrate_and_rename" {
+  name = "${var.basename}and-rename-lb"
   plan = "development"
   zone = var.zone